fix: split social into author & contributor
